refactor: add a create person controller and update files imports
